+++ src/index.ts
@@ -1,11 +1,13 @@
 import { createDataItemSigner } from "@permaweb/aoconnect";
-import { encrypt, decrypt, keygen, THRESHOLD_2_3} from "./algorithm";
+import { encrypt, decrypt, keygen, THRESHOLD_2_3 } from "./algorithm";
 import { nodes } from "./processes/noderegistry";
 import { register as dataRegister, getDataById } from "./processes/dataregistry";
-import { submit, getCompletedTasksById, /*getPendingTasks*/ } from "./processes/tasks";
+import { submit, getCompletedTasksById } from "./processes/tasks";
 import { submitDataToAR, getDataFromAR } from "./padoarweave";
+import { NODE_NAMES } from "./config";
+export { transferAOCREDToTask } from './processes/utils';
 import Arweave from 'arweave';
-import { readFileSync } from "node:fs";
+
 
 export interface PriceInfo {
   price: string;
@@ -13,62 +15,92 @@ export interface PriceInfo {
 }
 
 /**
- * Encrypt data and upload data
- *
- * @param data - plain data need to encrypt and upload
- * @param dataTag - the data meta info
- * @param priceInfo - The data price symbol and price
- * @param wallet - The ar wallet
- * @param arweave - The ar object and default is ar production
- * @returns The uploaded encrypted data id
- *
- */
+* Encrypt data and upload data
+*
+* @param data - plain data need to encrypt and upload
+* @param dataTag - the data meta info
+* @param priceInfo - The data price symbol and price
+* @param wallet - The ar wallet
+* @param arweave - The ar object and default is ar production
+* @returns The uploaded encrypted data id
+*/
 export const uploadData = async (data: Uint8Array, dataTag: any, priceInfo: PriceInfo,
   wallet: any, arweave: Arweave = Arweave.init({})): Promise<string> => {
-    if (data.length === 0) {
-      throw new Error("The Data to be uploaded can not be empty");
-    }
+  if (data.length === 0) {
+    throw new Error("The Data to be uploaded can not be empty");
+  }
 
-    priceInfo.symbol = priceInfo.symbol || "PADO Token";
+  priceInfo.symbol = priceInfo.symbol || "PADO Token";
 
-    let nodesres = await nodes();
-    nodesres = JSON.parse(nodesres);
-    let nodepks =Object();
-    for(let i in nodesres) {
-        let node = nodesres[i];
-        nodepks[node.name] = node.publickey;
-    }
-    let nodesPublicKey = [
-        nodepks["testnode1"],
-        nodepks["testnode2"],
-        nodepks["testnode3"],
-    ];
+  let nodesres = await nodes();
+  nodesres = JSON.parse(nodesres);
+  let nodepks = Object();
+  for (let i in nodesres) {
+    let node = nodesres[i];
+    nodepks[node.name] = node.publickey;
+  }
+  let nodesPublicKey = [];
+  for (let i in NODE_NAMES) {
+    nodesPublicKey.push(nodepks[NODE_NAMES[i]]);
+  }
 
-    const res = encrypt(nodesPublicKey, data);
+  const res = encrypt(nodesPublicKey, data);
 
-    const transactionId = await submitDataToAR(arweave, res.enc_msg, wallet);
+  const transactionId = await submitDataToAR(arweave, res.enc_msg, wallet);
 
-    const signer = createDataItemSigner(wallet);
-    const encSksStr = JSON.stringify(res.enc_sks);
-    const dataRes = await dataRegister(JSON.stringify(dataTag),
-      JSON.stringify(priceInfo), encSksStr, res.nonce, transactionId, signer);
-
-    return dataRes;
-}
-
-export const submitTask = async (dataId: string, dataUserPk: string, wallet: any): Promise<string> => {
   const signer = createDataItemSigner(wallet);
-  let inputData = {...THRESHOLD_2_3, dataId: dataId, consumerPk: dataUserPk};
-  const taskId = await submit("ZKLHEDataSharing", dataId, JSON.stringify(inputData),
-  "9000000000000", "512M",["testnode1", "testnode2", "testnode3"], signer);
-  return taskId;
+  const encSksStr = JSON.stringify(res.enc_sks);
+  // console.log('encSksStr', encSksStr);
+  // console.log('res.nonce', res.nonce);
+  const dataRes = await dataRegister(JSON.stringify(dataTag),
+    JSON.stringify(priceInfo), encSksStr, res.nonce, transactionId, signer);
+
+  // console.log('res.dataRes', dataRes);
+  return dataRes;
 }
 
 /**
- * Get the result of the task
- * @param taskId taskId
- * @param timeout Timeout in milliseconds
+ * Generate key pair for encrypt/decrypt
+ *
+ * @returns The key-pair object
  */
+export const generateKey = (): Promise<any> => {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve(keygen());
+    }, 1000);
+  });
+}
+
+/*
+export const listData = async () => {
+  // 1. get data list from data process
+  // 2. return dataTag, data ar url, data id, data price
+  // let datas = await allData();
+}*/
+
+
+//TODO:
+const taskType = "ZKLHEDataSharing";
+const computeLimit = "9000000000000";
+const memoryLimit = "512M";
+
+/**
+ * Submit a task to AO
+ *
+ * @param dataId - The data id
+ * @param dataUserPk - The user's public key generated by keygen
+ * @param wallet - The ar wallet
+ * @returns The submited task id
+ */
+export const submitTask = async (dataId: string, dataUserPk: string, wallet: any): Promise<string> => {
+  const signer = createDataItemSigner(wallet);
+  let inputData = { ...THRESHOLD_2_3, dataId: dataId, consumerPk: dataUserPk };
+  const taskId = await submit(taskType, dataId, JSON.stringify(inputData),
+    computeLimit, memoryLimit, NODE_NAMES, signer);
+  return taskId;
+}
+
 const getCompletedTaskPromise = (taskId: string, timeout: number): Promise<string> => {
   return new Promise((resolve, reject) => {
     const start = performance.now();
@@ -88,18 +120,24 @@ const getCompletedTaskPromise = (taskId: string, timeout: number): Promise<strin
   });
 };
 
+
+/**
+ * Get the result of the task
+ * 
+ * @param taskId The task id
+ * @param dataUserSk - The user's secret key generated by keygen
+ * @param arweave - The ar object and default is ar production
+ * @param timeout Timeout in milliseconds (default: 10 seconds)
+ * @returns The data
+ */
 export const getResult = async (taskId: string, dataUserSk: string,
-                                arweave: Arweave = Arweave.init({}), timeout?: number) => {
-  if(!timeout){
-    //default timeout is 10 seconds
-    timeout = 10000;
-  }
+  arweave: Arweave = Arweave.init({}), timeout: number = 10000) => {
   const taskStr = await getCompletedTaskPromise(taskId, timeout);
   const task = JSON.parse(taskStr);
   const chosenIndices = [1, 2];
   let reencSks = [];
   const computeNodes = JSON.parse(task.computeNodes);
-  for(let nodeName of computeNodes) {
+  for (let nodeName of computeNodes) {
     const reencSksObj = JSON.parse(task.result[nodeName]);
     reencSks.push(reencSksObj.reenc_sk);
   }
@@ -115,46 +153,8 @@ export const getResult = async (taskId: string, dataUserSk: string,
   return new Uint8Array(res.msg);
 };
 
-/*export const listData = async () => {
-    // 1. get data list from data process
-    // 2. return dataTag, data ar url, data id, data price
-}
 
+/*
 export const SubmitTaskAndGetResult = async (dataId: string, publicKey: string) => {
-    
-}*/
-
-async function test() {
-    const wallet = JSON.parse(
-        // readFileSync("/Users/fksyuan/Downloads/arweave-keyfile-JNqOSFDeSAh_icEDVAVa_r9wJfGU9AYCAJUQb2ss7T8.json").toString(),
-        readFileSync("/Users/xuda/script/wallets/ar/arweave-keyfile-gle8HJDn_QnEroXw1eb5cdCz9JMMwXn0ZynBPArHEQg.json").toString(),
-    );
-    //const signer = createDataItemSigner(wallet);
-    const arweave = Arweave.init({
-      host: '127.0.0.1',
-      port: 1984,
-      protocol: 'http'
-    });
-
-    const dataId = await uploadData(new Uint8Array([1,2,3,4,5,6]), {"testtagkey": "testtagvalue"}, {price: "1"}, wallet, arweave);
-    console.log("dataId=", dataId);
-    const dataResRes = await getDataById(dataId);
-    console.log("dataResRes=", dataResRes);
-
-    const dataUserKey = keygen();
-    //console.log("dataUserKey=", dataUserKey);
-    const taskId = await submitTask(dataId, dataUserKey.pk, wallet);
-    console.log("taskId=", taskId);
-    //const pendingTasks = await getPendingTasks();
-    //console.log("pendingTasks=", pendingTasks);*/
-
-    const res = await getResult(taskId, dataUserKey.sk, arweave);
-    console.log('res=', res);
-
-
-    /*setTimeout(async ()=> {
-      await register("testnode3", keygen().pk, "testnode3desc", createDataItemSigner(wallet));
-      await nodes();
-    }, 1000);*/
 }
-test();
+*/
